change workflow and file structure for tests
